+++ CHANGELOG.md
@@ -23,7 +23,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Update local docker-composer
 
 ### Added
+  - New Command `misc:register:command` to Register Command to service.yaml files.
   - ./bin/generate_service_yaml.php to generate `services.yaml` for `oe-console`
+  - ./bin/oxrun-light.php small application with one command `misc:register:command`
+
 
 ### Changed
   - Oxrun is now a [OXID eShop Component](https://docs.oxid-esales.com/developer/en/6.2/development/modules_components_themes/component.html)